server/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
ortools==9.10.4067
cachetools==5.3.3